requirements.txt
einops>=0.7.0
tqdm
mido
pretty_midi
numpy>=1.22.4
torchaudio>=0.11.0
onnxruntime-gpu>=1.19.0
madmom @ git+https://github.com/CPJKU/madmom
